requirements.txt
astroid==2.1.0
confusable-homoglyphs==3.2.0
coverage==4.5.2
Django==2.1.5
django-cors-headers==2.4.0
django-registration==3.0
djangorestframework==3.9.1
factory-boy==2.11.1
Faker==1.0.2
flake8==3.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
psycopg2-binary==2.7.7
pycodestyle==2.4.0
pydocstyle==3.0.0
pyflakes==2.0.0
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
snowballstemmer==1.2.1
text-unidecode==1.2
typed-ast==1.2.0
wrapt==1.11.1
